requirements.txt
accelerate
colorful==0.5.7
colorlog==6.10.1
diffusers==0.36.0
dill==0.4.0
einops==0.8.1
#flash_attn==2.8.2
huggingface-hub==0.36.0
multiprocess==0.70.16
numba
numpy
omegaconf==2.3.0
openai==1.106.1
packaging==25.0
pandas==2.3.3
peft==0.17.1
pillow==12.1.0
qwen-vl-utils==0.0.14
ray==2.50.0
timm==1.0.22
tokenizers==0.22.1
torch==2.8.0
torchaudio==2.8.0
torchdata==0.11.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.0
triton==3.4.0
typer==0.19.2
typer-slim==0.21.1
vllm==0.11.0
wandb==0.22.2
wheel==0.45.1
